Adapt benchmark to run in both macro and non-macro context
