+++ Cargo.toml
@@ -10,6 +10,7 @@ keywords = ["syn"]
 categories = ["development-tools::procedural-macro-helpers"]
 readme = "README.md"
 edition = "2018"
+autobenches = false
 rust-version = "1.31"
 
 [dependencies]
@@ -24,6 +25,9 @@ default = ["proc-macro"]
 # Disabling the proc-macro feature removes the dynamic library dependency on
 # libproc_macro in the rustc compiler.
 proc-macro = ["proc-macro2/proc-macro"]
+
+[workspace]
+members = ["benches"]
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
