+++ benches/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "quote-benchmark"
+version = "0.0.0"
+authors = ["David Tolnay <dtolnay@gmail.com>"]
+license = "MIT OR Apache-2.0"
+edition = "2018"
+publish = false
+
+[lib]
+proc-macro = true
+path = "lib.rs"
+
+[[bin]]
+name = "quote-benchmark"
+path = "main.rs"
+
+[dependencies]
+proc-macro2 = "1.0"
+quote = { path = ".." }
+termcolor = "1.1"
